Fix case where design matrix covariate has NaN
Problem: The spikes and design matrix didn’t have the same indices
causing the following error:

> ValueError: The indices for endog and exog are not aligned

This was caused by NaNs in the trajectory direction covariate.
Patsy.dmatrix automatically drops the NaNs in the design matrix
construction but this isn’t dropped in the response variable resulting
in the error. The solution here is to reindex the response dataframe to
match the design matrix. Apparently drop=‘missing’ in the GLM stats
models only applies to the response variable?