Change the default replay speed
According to  **Davidson, T.J., Kloosterman, F., and Wilson, M.A.
(2009). Hippocampal Replay of Extended Experience. Neuron 63,
497–507.**:

> … replay speeds are 15–20 times faster than a typical rat running
speed ( 0.5 m/s), consistent with previous reports of compression
factors for shorter-duration replay events (Lee and Wilson, 2002; Na ́
dasdy et al., 1999).

We had it at 30, so it makes sense to make the default more toward the
low end of the replay speeds. In the future, it might be good to
estimate the speed with EM (see issue #77)