Fix so that functions will look for raw data directory in the same place every time
Before the user was required to be one level up, i.e. in the notebooks
directory, for the correct directory to be returned. Because I always
used it from the notebooks directory, I did not notice this error. This
fix finds the directory of the data_processing module and expects the
root folder to be one level down. The Raw-Data folder and
Processed-Data folder are folders in the root folder.